agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Subarno",email = "subathanatos97@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.0,<2.0.0",
    "langchain>=0.1.0,<1.0.0",
    "langchain-openai>=0.1.0,<1.0.0",
    "langchain-core>=0.1.0,<1.0.0",
    "langchain-community>=0.0.20,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "fastapi>=0.104.0,<1.0.0",
    "uvicorn>=0.24.0,<1.0.0",
    "huggingface-hub>=0.34.0,<1.0.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
